# node-output/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "Genius Integration Services API"
host: "eugen-dev.r02.xlgs.local"
basePath: "/XLCapital.XLI.GISiStage2/Gapi"
schemes:
- "https"
paths:
...
  /api/risks/{riskreferencenumber}:
    get:
      tags:
      - "Risks"
      summary: "Get Risk Details when searching only by RiskReference."
      operationId: "Risk_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "riskreferencenumber"
        in: "path"
        description: "The risk reference number e.g. SE00000847LI17A"
        required: true
        type: "string"
      - name: "Authorization"
        in: "header"
        description: "Provide Bearer Token"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/RiskResponse"
      x-swagger-router-controller: "Risks"
  /api/risks/{riskreferencenumber}/coverages:
    get:
      tags:
      - "Risks"
      summary: "Get Coverage Details when searching only by RiskReference."
      operationId: "RiskCoverages_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "riskreferencenumber"
        in: "path"
        description: "The risk reference number e.g. SE00000847LI17A"
        required: true
        type: "string"
      - name: "Authorization"
        in: "header"
        description: "Provide Bearer Token"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/RiskCoveragesResponse"
      x-swagger-router-controller: "Risks"
  /api/risks/{riskreferencenumber}/name:
    put:
      tags:
      - "Risks"
      summary: "Add Name Involvement to the risk"
      description: "If you enter risk reference number along with name name and name\
        \ usage, then name gets added to the risk."
      operationId: "RiskNameInvolvements_Put"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "riskreferencenumber"
        in: "path"
        description: "The risk reference number e.g. AT00000292LI22A or QAT0001665LI22A"
        required: true
        type: "string"
      - in: "body"
        name: "request"
        description: "The data in json format details the fields to add name to the\
          \ risk."
        required: true
        schema:
          $ref: "#/definitions/NameInvolvementRequest"
      - name: "globalTrackingId"
        in: "header"
        description: "Provide Global Tracking Id"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/IGenericCommonResponse"
      x-swagger-router-controller: "Risks"
  /api/risks/{riskreferencenumber}/names:
    get:
      tags:
      - "Risks"
      summary: "Get Risk Details when searching only by RiskReference."
      operationId: "RiskNameInvolvements_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "riskreferencenumber"
        in: "path"
        description: "The risk reference number e.g. SE00000847LI17A"
        required: true
        type: "string"
      - name: "Authorization"
        in: "header"
        description: "Provide Bearer Token"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/RiskNamesResponse"
      x-swagger-router-controller: "Risks"
  /api/risks/{riskreferencenumber}/participations:
    get:
      tags:
      - "Risks"
      summary: "Get Risk Participations Details when searching only by RiskReference."
      operationId: "RiskParticipations_Get"
      consumes: []
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "riskreferencenumber"
        in: "path"
        description: "The risk reference number e.g. SE00000847LI17A"
        required: true
        type: "string"
      - name: "Authorization"
        in: "header"
        description: "Provide Bearer Token"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/RiskParticipationsResponse"
      x-swagger-router-controller: "Risks"
  /api/risks/{riskreferencenumber}/status:
    put:
      tags:
      - "Risks"
      summary: "Update the Genius Policy Master Status Code on a Local Policy"
      description: "If you enter policy reference number along with policy master\
        \ status code and policy narrative, then policy status gets updated."
      operationId: "RisksStatus_Put"
      consumes:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      - "text/json"
      - "application/xml"
      - "text/xml"
      parameters:
      - name: "riskreferencenumber"
        in: "path"
        description: "The policy reference number e.g. AT00015092PR18A"
        required: true
        type: "string"
      - in: "body"
        name: "request"
        description: "The data in json format details the fields required to update\
          \ the policy status."
        required: true
        schema:
          $ref: "#/definitions/UpdatePolicyStatusRequest"
      - name: "globalTrackingId"
        in: "header"
        description: "Provide Global Tracking Id"
        required: true
        type: "string"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/UpdatePolicyStatusResponse"
      x-swagger-router-controller: "Risks"
definitions:
...
  StatusMessage:
    type: "object"
    properties:
      MessageType:
        type: "string"
        enum:
        - "Error"
        - "Warning"
        - "Information"
      MessageCode:
        type: "string"
      MessageDescription:
        type: "string"
      MessageDetail:
        type: "string"
    example:
      MessageCode: "MessageCode"
      MessageDescription: "MessageDescription"
      MessageDetail: "MessageDetail"
      MessageType: "Error"
...
  IGenericCommonResponse:
    type: "object"
...
  RiskResponse:
    type: "object"
    properties:
      riskDetails:
        $ref: "#/definitions/RiskMaster"
      _links:
        type: "array"
        items:
          $ref: "#/definitions/RedirectLinks"
      total:
        type: "integer"
        format: "int32"
    example:
      total: 5
      _links:
      - rel: "rel"
        href: "href"
      - rel: "rel"
        href: "href"
      riskDetails:
        riskTitle: "riskTitle"
        lineOfBusiness: "lineOfBusiness"
        programAccountName: "programAccountName"
        tax: 1.4658129805029452
        programReferenceNumber: "programReferenceNumber"
        premiumType: "premiumType"
        expiryDate: "2000-01-23T04:56:07.000+00:00"
        currencyOfAmounts: "currencyOfAmounts"
        productCode: "productCode"
        instalmentFrequency: "instalmentFrequency"
        programAccountNumber: "programAccountNumber"
        inceptionDate: "2000-01-23T04:56:07.000+00:00"
        policyReference: "policyReference"
        grossPremium: 6.027456183070403
        invoiceNumber: "invoiceNumber"
        quoteReference: "quoteReference"
        underwritingYear: 0
        currency: "currency"
        underWritingLocation: "underWritingLocation"
        masterStatusCode: "masterStatusCode"
  RiskMaster:
    type: "object"
    properties:
      quoteReference:
        type: "string"
      policyReference:
        type: "string"
      programReferenceNumber:
        type: "string"
      riskTitle:
        type: "string"
      masterStatusCode:
        type: "string"
      inceptionDate:
        type: "string"
        format: "date-time"
      underwritingYear:
        type: "integer"
        format: "int32"
      expiryDate:
        type: "string"
        format: "date-time"
      lineOfBusiness:
        type: "string"
      productCode:
        type: "string"
      programAccountName:
        type: "string"
      programAccountNumber:
        type: "string"
      underWritingLocation:
        type: "string"
      currencyOfAmounts:
        type: "string"
      currency:
        type: "string"
      premiumType:
        type: "string"
      grossPremium:
        type: "number"
        format: "double"
      tax:
        type: "number"
        format: "double"
      instalmentFrequency:
        type: "string"
      invoiceNumber:
        type: "string"
    example:
      riskTitle: "riskTitle"
      lineOfBusiness: "lineOfBusiness"
      programAccountName: "programAccountName"
      tax: 1.4658129805029452
      programReferenceNumber: "programReferenceNumber"
      premiumType: "premiumType"
      expiryDate: "2000-01-23T04:56:07.000+00:00"
      currencyOfAmounts: "currencyOfAmounts"
      productCode: "productCode"
      instalmentFrequency: "instalmentFrequency"
      programAccountNumber: "programAccountNumber"
      inceptionDate: "2000-01-23T04:56:07.000+00:00"
      policyReference: "policyReference"
      grossPremium: 6.027456183070403
      invoiceNumber: "invoiceNumber"
      quoteReference: "quoteReference"
      underwritingYear: 0
      currency: "currency"
      underWritingLocation: "underWritingLocation"
      masterStatusCode: "masterStatusCode"
  RedirectLinks:
    type: "object"
    properties:
      href:
        type: "string"
        readOnly: true
      rel:
        type: "string"
        readOnly: true
    example:
      rel: "rel"
      href: "href"
  RiskCoveragesResponse:
    type: "object"
    properties:
      coverages:
        type: "array"
        items:
          $ref: "#/definitions/RiskCoverages"
      total:
        type: "integer"
        format: "int32"
    example:
      total: 5
      coverages:
      - grosspremium: 5.962133916683182
        coverageName: "coverageName"
        geniusKZTaxFlag: "geniusKZTaxFlag"
        insuranceTaxPercentage: 6.027456183070403
        section: "section"
        subSection: "subSection"
        countryName: "countryName"
        sectionDetail: "sectionDetail"
        coverageCode: "coverageCode"
        netPremium: 0.8008281904610115
        insuranceTax: 1.4658129805029452
      - grosspremium: 5.962133916683182
        coverageName: "coverageName"
        geniusKZTaxFlag: "geniusKZTaxFlag"
        insuranceTaxPercentage: 6.027456183070403
        section: "section"
        subSection: "subSection"
        countryName: "countryName"
        sectionDetail: "sectionDetail"
        coverageCode: "coverageCode"
        netPremium: 0.8008281904610115
        insuranceTax: 1.4658129805029452
  RiskCoverages:
    type: "object"
    properties:
      section:
        type: "string"
      subSection:
        type: "string"
      sectionDetail:
        type: "string"
      coverageCode:
        type: "string"
      coverageName:
        type: "string"
      countryName:
        type: "string"
      netPremium:
        type: "number"
        format: "double"
      insuranceTaxPercentage:
        type: "number"
        format: "double"
      insuranceTax:
        type: "number"
        format: "double"
      grosspremium:
        type: "number"
        format: "double"
      geniusKZTaxFlag:
        type: "string"
    example:
      grosspremium: 5.962133916683182
      coverageName: "coverageName"
      geniusKZTaxFlag: "geniusKZTaxFlag"
      insuranceTaxPercentage: 6.027456183070403
      section: "section"
      subSection: "subSection"
      countryName: "countryName"
      sectionDetail: "sectionDetail"
      coverageCode: "coverageCode"
      netPremium: 0.8008281904610115
      insuranceTax: 1.4658129805029452
  NameInvolvementRequest:
    type: "object"
    properties:
      nameCode:
        type: "string"
      nameType:
        type: "string"
      requestUser:
        type: "string"
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      nameCode: "ABARKER"
      nameType: "DAM"
      requestUser: "UserID"
  RiskNamesResponse:
    type: "object"
    properties:
      names:
        type: "array"
        items:
          $ref: "#/definitions/RiskNames"
      total:
        type: "integer"
        format: "int32"
    example:
      total: 0
      names:
      - nameType: "nameType"
        areaCode: "areaCode"
        nameCode: "nameCode"
        name: "name"
        xlNameID: "xlNameID"
        addressLine1: "addressLine1"
        addressLine2: "addressLine2"
        nameUsage: "nameUsage"
        addressLine3: "addressLine3"
        addressLine4: "addressLine4"
        dunsNumber: "dunsNumber"
        addressLine5: "addressLine5"
      - nameType: "nameType"
        areaCode: "areaCode"
        nameCode: "nameCode"
        name: "name"
        xlNameID: "xlNameID"
        addressLine1: "addressLine1"
        addressLine2: "addressLine2"
        nameUsage: "nameUsage"
        addressLine3: "addressLine3"
        addressLine4: "addressLine4"
        dunsNumber: "dunsNumber"
        addressLine5: "addressLine5"
  RiskNames:
    type: "object"
    properties:
      nameType:
        type: "string"
      nameUsage:
        type: "string"
      nameCode:
        type: "string"
      name:
        type: "string"
      dunsNumber:
        type: "string"
      xlNameID:
        type: "string"
      addressLine1:
        type: "string"
      addressLine2:
        type: "string"
      addressLine3:
        type: "string"
      addressLine4:
        type: "string"
      addressLine5:
        type: "string"
      areaCode:
        type: "string"
    example:
      nameType: "nameType"
      areaCode: "areaCode"
      nameCode: "nameCode"
      name: "name"
      xlNameID: "xlNameID"
      addressLine1: "addressLine1"
      addressLine2: "addressLine2"
      nameUsage: "nameUsage"
      addressLine3: "addressLine3"
      addressLine4: "addressLine4"
      dunsNumber: "dunsNumber"
      addressLine5: "addressLine5"
  RiskParticipationsResponse:
    type: "object"
    properties:
      participations:
        type: "array"
        items:
          $ref: "#/definitions/RiskParticipations"
      total:
        type: "integer"
        format: "int32"
    example:
      participations:
      - participantNameCode: "participantNameCode"
        portion: 0.8008281904610115
        grossPremium: 5.962133916683182
        insuranceTaxNo: "insuranceTaxNo"
        participantName: "participantName"
        netPremium: 6.027456183070403
        insuranceTax: 1.4658129805029452
      - participantNameCode: "participantNameCode"
        portion: 0.8008281904610115
        grossPremium: 5.962133916683182
        insuranceTaxNo: "insuranceTaxNo"
        participantName: "participantName"
        netPremium: 6.027456183070403
        insuranceTax: 1.4658129805029452
      total: 5
  RiskParticipations:
    type: "object"
    properties:
      participantNameCode:
        type: "string"
      participantName:
        type: "string"
      insuranceTaxNo:
        type: "string"
      portion:
        type: "number"
        format: "double"
      netPremium:
        type: "number"
        format: "double"
      insuranceTax:
        type: "number"
        format: "double"
      grossPremium:
        type: "number"
        format: "double"
    example:
      participantNameCode: "participantNameCode"
      portion: 0.8008281904610115
      grossPremium: 5.962133916683182
      insuranceTaxNo: "insuranceTaxNo"
      participantName: "participantName"
      netPremium: 6.027456183070403
      insuranceTax: 1.4658129805029452
  UpdatePolicyStatusRequest:
    type: "object"
    properties:
      PolicyReference:
        type: "string"
      Status:
        type: "string"
      Narrative:
        type: "string"
      Reason:
        type: "string"
      RequestUser:
        type: "string"
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      PolicyReference: "AT00015092PR18A"
      Status: "NF"
      Narrative: "Adding Policy Narrative"
      Reason: ""
      RequestUser: "UserID"
  UpdatePolicyStatusResponse:
    type: "object"
    properties:
      PolicyReference:
        type: "string"
      RequestUser:
        type: "string"
      Reference:
        type: "string"
      GeniusTransactionId:
        type: "integer"
        format: "int32"
      Messages:
        type: "array"
        items:
          type: "string"
      StatusMessages:
        type: "array"
        items:
          $ref: "#/definitions/StatusMessage"
      ProcessStatus:
        type: "integer"
        format: "int32"
        enum:
        - 0
        - 1
        - 2
      GlobalTrackingId:
        type: "string"
      SourceSystem:
        type: "string"
    example:
      GeniusTransactionId: 0
      Messages:
      - "Messages"
      - "Messages"
      ProcessStatus: 6
      Reference: "Reference"
      GlobalTrackingId: "GlobalTrackingId"
      StatusMessages:
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      - MessageCode: "MessageCode"
        MessageDescription: "MessageDescription"
        MessageDetail: "MessageDetail"
        MessageType: "Error"
      PolicyReference: "PolicyReference"
      RequestUser: "RequestUser"
      SourceSystem: "SourceSystem"